Linked player health to AI impact and destruction
Created the logic in the Blueprints of both the player and the AI to link the destruction of the AI to health decrements of the player. Also created the logic to measure the player's health at any given moment and continuously calculate how much health is remaining before reaching zero.