Fixed problem where AI does not sense player
Fixed a problem in Blueprints where the AI does not sense the player. This bug occurs when using Blueprints, so to fix it I added a nav Mesh Bounds Volume box to the scene and set the size to match the scene. As long as the player and the AI are within this box the AI will detect the player and begin homing in.