Completed the final environment enclosure play space
I have created a series of laser bars and used these to enclose the playable area. The playable environment is now complete and ready for detailing and gameplay dynamics. Best of all, only one material (with zero textures) has been used and repeated with varied instances to create the complete look and feel for the scene. The final landscape looks amazing, with no loss or diversion of resources to heavy textures.